requirements.txt
autogen
openai
faiss-cpu
sentence-transformers
python-dotenv